requirements.txt
langchain==0.3.7
streamlit==1.39.0
langchain-community==0.3.5
openai==1.54.1
langchain-openai==0.2.6
streamlit-chat==0.1.1
pypdf==5.1.0
faiss-cpu==1.9.0
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
openai-whisper==20240930
transformers==4.46.2
pydub==0.25.1